REF: Move create_exceptions_mapping into a `utils` module
